pyproject.toml
[project]
name = "demo-polyu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bosdyn-choreography-client>=5.0.1.2",
    "bosdyn-client>=5.0.1.2",
    "bosdyn-mission>=5.0.1.2",
    "bosdyn-orbit>=5.0.1.2",
    "datasets>=4.5.0",
    "ffmpeg>=1.4",
    "flask>=3.1.2",
    "numpy>=2.3.4",
    "open3d>=0.19.0",
    "opencv-python>=4.11.0.86",
    "openvino>=2026.0.0",
    "pynput>=1.8.1",
    "ultralytics>=8.3.225",
    "vtk>=9.5.2",
]
